Basic User Interface
+Fixed Videogame class, GetYearOfRelease actually returns an int now,
SetYearOfRelease actually converts properly to int

+Added Basic user interface, allows for window resize and dynamic
scaling
+Basic Add and Delete slots added
